Address PR #1104 review feedback
- Fix stale module-level storage structure diagram (issue 1)
- Fix incorrect path in migration and symlink doc comments (issues 2, 3)
- Repair broken symlinks when target matches but doesn't exist (issue 4)
- Handle directories in stale link removal, not just files (issue 5)
- Move unit tests to separate auth_tests.rs file (issue 6)
- Update README.md with new auth directory structure (issue 7)
- Add tests for broken symlink repair and directory replacement
